update files
remove unused files

rename to use right python conventions

update imports elsewhere
