--- Magias/dfs/global_vars.json
@@ -1,78 +1,0 @@
-{
-  "G_ESCOLAS": [
-    "elemental",
-    "necromancia",
-    "psíquica",
-    "ilusionista",
-    "invocação",
-    "espiritual",
-    "atrativa",
-    "musical",
-    "pura",
-    "receita"
-  ],
-  "G_ELEMENTOS": [
-    "ar",
-    "fogo",
-    "luz",
-    "metal",
-    "relâmpago",
-    "sombras",
-    "terra",
-    "veneno",
-    "água"
-  ],
-  "G_TAGS": [
-    "area",
-    "buff",
-    "debuff",
-    "controle",
-    "utilidade",
-    "dano",
-    "defesa",
-    "cura",
-    "distância",
-    "corpo-a-corpo",
-    "toque",
-    "arma",
-    "comunicação",
-    "social",
-    "detecção",
-    "exploração"
-  ],
-  "G_CLASSES": [
-    "arqueiro",
-    "bardo",
-    "monge",
-    "clérigo",
-    "paladino",
-    "ladino",
-    "guerreiro",
-    "mago",
-    "bruxo",
-    "feiticeiro",
-    "xamã"
-  ],
-  "G_COLUMNS": [
-    "nome",
-    "name",
-    "nivel",
-    "escola",
-    "ritual",
-    "elementos",
-    "tempo_conjuracao",
-    "alcance_area",
-    "componentes",
-    "mana",
-    "duracao",
-    "attack_save",
-    "dmg_effect",
-    "dmg",
-    "classes",
-    "tags",
-    "descricao",
-    "source",
-    "mana_adicional",
-    "magia_rara"
-  ]
-}
