+++ Magias/dfs/schema_config.json
@@ -65,8 +65,5 @@
         "dmg": "N/A",
         "mana_adicional": "N/A",
         "magia_rara": false
-    },
-    "column_possible_values": {
-
     }
 }
